# sdks/sdkwork-search-backend-sdk/openapi/sdkwork-search-backend-api.openapi.yaml
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Search Backend API",
    "version": "1.0.0",
    "description": "Backend/admin contract for sdkwork-search index and document management APIs.",
    "x-sdkwork-api-authority": "sdkwork-search-backend-api",
    "x-sdkwork-sdk-family": "sdkwork-search-backend-sdk",
    "x-sdkwork-audience": "Backend console, operators, control plane, and admin integrations"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local sdkwork-search runtime"
    }
  ],
...
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
    "/backend/v3/api/search/indexes": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "List search indexes for backend administration.",
        "operationId": "search.indexes.list",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Standard free-text search query."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchIndexListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-search",
        "x-sdkwork-api-authority": "sdkwork-search-backend-api",
        "x-sdkwork-domain": "search",
        "x-sdkwork-resource": "search.indexes",
        "x-sdkwork-permission": "search.indexes.read",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "search.index.list",
        "x-sdkwork-idempotent": true,
        "x-sdkwork-deployment": "all",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-source": "packages/native-rust/routes/backend-api/sdkwork-routes-search-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-search-backend-api",
        "x-sdkwork-server-request-id": true
      },
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Create a search index.",
        "operationId": "search.indexes.create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            },
            "description": "Client retry key for retriable create commands."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchIndexResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-search",
        "x-sdkwork-api-authority": "sdkwork-search-backend-api",
        "x-sdkwork-domain": "search",
        "x-sdkwork-resource": "search.indexes",
        "x-sdkwork-permission": "search.indexes.write",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "search.index.create",
        "x-sdkwork-idempotent": true,
        "x-sdkwork-deployment": "all",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-source": "packages/native-rust/routes/backend-api/sdkwork-routes-search-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-search-backend-api",
        "x-sdkwork-server-request-id": true,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchIndexCreateRequest"
              }
            }
          }
        }
      }
    },
    "/backend/v3/api/search/indexes/{indexId}": {
      "patch": {
        "tags": [
          "search"
        ],
        "summary": "Update a search index.",
        "operationId": "search.indexes.update",
        "parameters": [
          {
            "name": "indexId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchIndexResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-search",
        "x-sdkwork-api-authority": "sdkwork-search-backend-api",
        "x-sdkwork-domain": "search",
        "x-sdkwork-resource": "search.indexes",
        "x-sdkwork-permission": "search.indexes.write",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "search.index.update",
        "x-sdkwork-idempotent": true,
        "x-sdkwork-deployment": "all",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-source": "packages/native-rust/routes/backend-api/sdkwork-routes-search-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-search-backend-api",
        "x-sdkwork-server-request-id": true,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchIndexUpdateRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "search"
        ],
        "summary": "Delete a search index.",
        "operationId": "search.indexes.delete",
        "parameters": [
          {
            "name": "indexId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchIndexDeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-search",
        "x-sdkwork-api-authority": "sdkwork-search-backend-api",
        "x-sdkwork-domain": "search",
        "x-sdkwork-resource": "search.indexes",
        "x-sdkwork-permission": "search.indexes.write",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "search.index.delete",
        "x-sdkwork-idempotent": true,
        "x-sdkwork-deployment": "all",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-source": "packages/native-rust/routes/backend-api/sdkwork-routes-search-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-search-backend-api",
        "x-sdkwork-server-request-id": true
      }
    },
    "/backend/v3/api/search/indexes/{indexId}/rebuild": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Rebuild a search index by dropping and recreating it.",
        "operationId": "search.indexes.rebuild",
        "parameters": [
          {
            "name": "indexId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchIndexJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-search",
        "x-sdkwork-api-authority": "sdkwork-search-backend-api",
        "x-sdkwork-domain": "search",
        "x-sdkwork-resource": "search.indexes",
        "x-sdkwork-permission": "search.indexes.write",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "search.index.rebuild",
        "x-sdkwork-idempotent": false,
        "x-sdkwork-deployment": "all",
        "x-sdkwork-request-context": "BackendRequestContext",
        "x-sdkwork-source": "packages/native-rust/routes/backend-api/sdkwork-routes-search-backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-search-backend-api",
        "x-sdkwork-server-request-id": true,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "schema"
                ],
                "properties": {
                  "schema": {
                    "type": "object",
                    "description": "Index schema used to recreate the index after drop."
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "SearchIndexListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "pageInfo",
          "requestId"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchIndex"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/SearchPageInfo"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
      "SearchIndexCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "indexId",
          "name"
        ],
        "properties": {
          "indexId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "analyzer": {
            "type": "string"
          }
        }
      },
      "SearchIndexResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "index",
          "requestId"
        ],
        "properties": {
          "index": {
            "$ref": "#/components/schemas/SearchIndex"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
      "SearchIndexUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "analyzer": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "archived",
              "draft",
              "paused"
            ]
          }
        }
      },
      "SearchIndexDeleteResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "deleted",
          "indexId",
          "requestId"
        ],
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "indexId": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SearchIndexJobResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jobId",
          "requestId",
          "status"
        ],
        "properties": {
          "jobId": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "status": {
            "type": "string",
            "enum": [
              "failed",
              "pending",
              "running",
              "succeeded"
            ]
          }
        }
      },
...
      "SearchIndex": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "indexId",
          "name",
          "documentCount",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "indexId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "analyzer": {
            "type": "string"
          },
          "documentCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}